fixed weird namings, restructured /students, fixed useStudents hook. removed unnecessary files
