+++ next.config.js
@@ -2,12 +2,12 @@
 const nextConfig = {
     webpack(config) {
         config.module.rules.push({
-          test: /\.svg$/,
-          use: ["@svgr/webpack"]
-        });
-    
-        return config;
-    }
+            test: /\.svg$/,
+            use: ["@svgr/webpack"],
+        })
+
+        return config
+    },
 }
 
 module.exports = nextConfig
