--- src/_temp_types/courseSections.ts
@@ -1,5 +1,0 @@
-import { Student } from "@/_temp_types/student"
-export type CourseSection = {
-    id: number
-    name?: String
-}
