--- src/_temp_types/courseStudents.ts
@@ -1,7 +1,0 @@
-import { Student } from "@/_temp_types/student"
-
-export type CourseStudents = {
-    id: number
-    name?: string
-    students: Student[]
-}
