--- src/types/CourseSection.ts
@@ -1,5 +1,0 @@
-import { Student } from "./Student"
-export type CourseSection = {
-    sectionId: number
-    sectionName?: String
-}
