--- src/types/CourseStudents.ts
@@ -1,7 +1,0 @@
-import {Student} from '@/_temp_types/student'
-
-export type CourseStudents = {
-    courseID: number,
-    courseName?: string,
-    students: Student[]
-}
